projects/KaymakanMaxim/PERSONAL/skills/telegram-telethon/requirements.txt
telethon>=1.36,<2.0
python-dotenv>=1.0
cryptg>=0.4         # C-accelerated crypto, big speedup for media
python-socks[asyncio]>=2.4  # for socks proxy support
